Vulkan backend
Rewrites the rendering code to use a common low-level interface instead of using OpenGL directly. The low-level renderer has two backends: OpenGL and Vulkan.

Adds a rendering test suite to verify that the low-level renderer works the same regardless of backend.

Removes particles as the existing system was not satisfactory and will have to be rewritten from scratch. Also removes the GPU profiling as it was heavily tied to OpenGL.