AMD bug reproducer
Adds a test which reproduces what is presumably a caching issue encountered on AMD Radeon RX 5700 XT.

The test writes a value to an SSBO, then reads it back in a fragment shader. It writes a second value to the SSBO and reads it back once again. The second readback receives an incorrect value (the first value rather than the second).